+++ src/Controlador/domicilioControlador.java
@@ -1,0 +1,24 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package Controlador;
+
+/**
+ *
+ * @author maxi
+ */
+public class domicilioControlador {
+    
+     public void agregar() {
+  }
+
+  public void actualizar() {
+  }
+
+  public void eliminar() {
+  }
+
+  public void mostrar() {
+  }
+}
