+++ src/Controlador/fichaKinesicaControlador.java
@@ -1,0 +1,23 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package Controlador;
+
+/**
+ *
+ * @author maxi
+ */
+public class fichaKinesicaControlador {
+     public void agregar() {
+  }
+
+  public void actualizar() {
+  }
+
+  public void eliminar() {
+  }
+
+  public void mostrar() {
+  }
+}
