+++ src/Modelo/Diagnostico.java
@@ -58,16 +58,6 @@ public class Diagnostico extends FichaKinesica {
         this.estado = estado;
     }
 
-  public void agregar() {
-  }
-
-  public void actualizar() {
-  }
-
-  public void mostrar() {
-  }
-
-  public void eliminar() {
-  }
+ 
 
 }
