+++ src/Modelo/Domicilio.java
@@ -59,16 +59,6 @@ public class Domicilio {
     }
 
 
-  public void agregar() {
-  }
-
-  public void actualizar() {
-  }
-
-  public void eliminar() {
-  }
-
-  public void mostrar() {
-  }
+ 
 
 }
